Started EditMod and RemoveMod Functions
Started the basics of the Edit and Remove Mod functions. The first thing both need to do is check to see if the mod that is to be edited or removed exists.

Also made the DisplayAllMods function, which just calls the DisplayMod function for every mod in the ModList vector.

Also made the FindMod function more generic by adding a SSEMod vector input, thus allowing the function to also work when I want to use it on the Dependencies vector that each mod has.