Dependency Arrays are now arrays of pointers
I THINK that I have been able to change the code so that the mDependencies array of the SSEMod struct is now an array of SSEMod pointers, rather than SSEMod instances. I will need to do some testing and probably get some help to see if everything was set up correctly.