Switched from Vectors in the Struct to a Map
Rather than having a vector of mods, ModList is now a map, with the key being a single mod, and the value being an array of pointers to mods, which will act as the dependencies. I think everything works as intended, but I do think I'll need a second pair of eyes on it to see if I missed anything.