Save and Load Progress
Saving and loading now work. However, I need to modify the save function. Right now, I don't know if it will save copies of mods that already exist. So, I need to read through the save file and, if there is already data with the mod's information, to not write to the file again. I also need to do a check to see if any mods were removed, and if they were, remove them from the file as well.