Working on Saving
The save function now clears the ModList.txt file and then saves the new information, so this does solve all of the potential problems I have. But I do not like the solution. I would like to see if there is a more elegant solution that does not involve rewriting all the information every time it is saved.